+++ grails-app/controllers/com/mini/asaas/customer/CustomerController.groovy
@@ -1,6 +1,6 @@
 package com.mini.asaas.customer
 
-import com.mini.asaas.AddressService
+import com.mini.asaas.adress.AddressService
 import com.mini.asaas.customer.Customer
 import com.mini.asaas.dto.customer.CustomerDTO
 import grails.compiler.GrailsCompileStatic
