+++ grails-app/services/com/mini/asaas/payer/PayerService.groovy
@@ -1,9 +1,9 @@
 package com.mini.asaas.payer
 
+import com.mini.asaas.adress.Address
+import com.mini.asaas.adress.AddressService
 import com.mini.asaas.payer.Payer
-import com.mini.asaas.Address
 import com.mini.asaas.customer.Customer
-import com.mini.asaas.AddressService
 import com.mini.asaas.dto.payer.PayerDTO
 import com.mini.asaas.utils.CpfCnpjUtils
