+++ src/main/groovy/com/mini/asaas/base/PersonEntity.groovy
@@ -1,6 +1,6 @@
 package com.mini.asaas.base
 
-import com.mini.asaas.Address
+import com.mini.asaas.adress.Address
 import com.mini.asaas.enums.PersonType
 import com.mini.asaas.base.BaseEntity
 import grails.compiler.GrailsCompileStatic
